pull routing and models from app.js to their own modules
