+++ public/js/edittool-lib.js
