+++ routes/edittool.js
@@ -1,0 +1,4 @@
+const express = require('express');
+const Tool = require('../models/tool.model.js');
+
+const router = express.Router();
